requirements.txt
crewai==0.80.0
crewai-tools==0.14.0
langchain==0.3.7
langchain-openai==0.2.9
langchain-community==0.3.7
langchain-groq==0.2.1
python-dotenv==1.0.1
pydantic==2.9.2
openai==1.54.4
groq==0.12.0
requests==2.32.3
PyYAML==6.0.2
rich==13.9.4
typer==0.13.0
